+++ e2e/client/option.go
@@ -163,7 +163,6 @@ func (o *TestTaskOptions) GetNewTask(ks *KeyStore) *asset.NewComputeTask {
 		Key:            ks.GetKey(o.KeyRef),
 		Category:       asset.ComputeTaskCategory_TASK_TEST,
 		AlgoKey:        ks.GetKey(o.AlgoRef),
-		ParentTaskKeys: parentKeys,
 		ComputePlanKey: ks.GetKey(o.PlanRef),
 		Data: &asset.NewComputeTask_Test{
 			Test: &asset.NewTestTaskData{
@@ -245,7 +244,6 @@ func (o *TrainTaskOptions) GetNewTask(ks *KeyStore) *asset.NewComputeTask {
 		Key:            ks.GetKey(o.KeyRef),
 		Category:       asset.ComputeTaskCategory_TASK_TRAIN,
 		AlgoKey:        ks.GetKey(o.AlgoRef),
-		ParentTaskKeys: parentKeys,
 		ComputePlanKey: ks.GetKey(o.PlanRef),
 		Data: &asset.NewComputeTask_Train{
 			Train: &asset.NewTrainTaskData{
@@ -312,7 +310,6 @@ func (o *PredictTaskOptions) GetNewTask(ks *KeyStore) *asset.NewComputeTask {
 		Key:            ks.GetKey(o.KeyRef),
 		Category:       asset.ComputeTaskCategory_TASK_PREDICT,
 		AlgoKey:        ks.GetKey(o.AlgoRef),
-		ParentTaskKeys: parentKeys,
 		ComputePlanKey: ks.GetKey(o.PlanRef),
 		Data: &asset.NewComputeTask_Predict{
 			Predict: &asset.NewPredictTaskData{
@@ -374,7 +371,6 @@ func (o *CompositeTaskOptions) GetNewTask(ks *KeyStore) *asset.NewComputeTask {
 		Key:            ks.GetKey(o.KeyRef),
 		Category:       asset.ComputeTaskCategory_TASK_COMPOSITE,
 		AlgoKey:        ks.GetKey(o.AlgoRef),
-		ParentTaskKeys: parentKeys,
 		ComputePlanKey: ks.GetKey(o.PlanRef),
 		Data: &asset.NewComputeTask_Composite{
 			Composite: &asset.NewCompositeTrainTaskData{
@@ -430,7 +426,6 @@ func (o *AggregateTaskOptions) GetNewTask(ks *KeyStore) *asset.NewComputeTask {
 		Key:            ks.GetKey(o.KeyRef),
 		Category:       asset.ComputeTaskCategory_TASK_AGGREGATE,
 		AlgoKey:        ks.GetKey(o.AlgoRef),
-		ParentTaskKeys: parentKeys,
 		ComputePlanKey: ks.GetKey(o.PlanRef),
 		Data: &asset.NewComputeTask_Aggregate{
 			Aggregate: &asset.NewAggregateTrainTaskData{