+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ModelService.GetComputeTaskInputModels, use ComputeTaskAPI.GetInputAssets instead
 - Test task rank special case: rank is not inherited from parent task anymore
 
+### Deprecated
+
+- `NewComputeTask.parent_task_keys` is deprecated since parent tasks are determined from task inputs
+
 ## [0.25.0] - 2022-08-29
 
 ### Added